Refactor to new file structure, update README.md
