+++ README.md
@@ -1,30 +1,22 @@
 # Rails Generator
 
-This is a CLI application used for creating a rails project according to the rails project setup guide. It also contains a small CLI utility for configuring an existing project.
+This is a CLI application used for creating a rails project according to the rails project setup guide. It also contains
+a small CLI utility for configuring an existing project.
 
 ## Prerequisites
 
 - Bun 0.8.1
 
-
 ## Installation
 
 1. Clone this repository `git clone git@github.com:renuo/rails-generator.git`
 2. Change into the project directory: `cd rails-generator`
-3. Install the dependencies: `npm install`
-4. Add alias to the application file in your `.zshrc` or `.bashrc` `alias grails=PATH_TO_INDEX.ts`
-5. Add alias to the cli tool in your `.zshrc` or `.bashrc` `alias crails=PATH_TO_PROGRAM.ts`
+3. Install the dependencies: `bun install`
+4. Add the `bin` directory to your `PATH` variable in your `.zshrc` or `.bashrc` `export PATH=$PATH:PATH_TO_CRAILS/bin`
 
 ## Usage
 
-### Creating a new application
-
-1. `cd` into the directory you the rails project to be created at
-2. Run `grails`
-
-### Configuring a existing application
-
-1. `cd` into the project directory
+1. `cd` into the directory you the rails project to be created at / configured in
 2. Run `crails`
 
 ## License
