+++ bin/crails
@@ -1,0 +1,3 @@
+#!/usr/bin/env bun
+
+require('../src/index.ts')
