+++ package.json
@@ -6,12 +6,12 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint --fix",
-    "build": "npx tsc"
+    "crails": "bin/crails"
   },
   "author": "CuddlyBunion341",
   "license": "MIT",
   "bin": {
-    "run": "./bin/index.js"
+    "crails": "./bin/crails"
   },
   "dependencies": {
     "@types/figlet": "^1.5.6",
