+++ src/index.ts
@@ -1,35 +1,29 @@
-#!/usr/bin/env bun
-
 /**
- * @file Runs the CLI application for configuring a Rails application
- * @description It is a straightforward application that utilizes the inquirer library to prompt the user
- * with a series of questions. Based on the answers provided,
- * it will generate a Rails application.
+ * @file Creates a new commander program for configuring a Rails application
+ * @description This is the entry point of the CLI tool
  * @author CuddlyBunion341
  */
 
+import { program } from 'commander'
+import figlet from 'figlet'
 import chalk from 'chalk'
-import inquirer from 'inquirer'
 
-const { log } = console // Destructure log, because we use it a lot
+import initCommand from './commands/init/init'
+import createCommand from './commands/create/create'
 
-/**
- * Main entry point for the CLI application
- */
-async function main (): Promise<void> {
-  log(chalk.bold('Hello World!'))
-
-  const projectName = await inquirer.prompt([
-    {
-      type: 'input',
-      name: 'name',
-      message: 'What is the name of your project?'
-    }
-  ])
-
-  log(chalk.bold(`TODO: Create a new Rails project with the name ${projectName.name}`))
+if (process.argv.length < 3) {
+  const asciiArt = figlet.textSync('crails', {
+    font: 'ANSI Shadow'
+  })
+  console.log(chalk.red(asciiArt))
 }
 
-main()
-  .then(() => { log(chalk.bold.green('Done!')) })
-  .catch((err) => { log(chalk.bold.red(err)) })
+program
+  .name('rails-configurator')
+  .description('A CLI tool for configuring Rails applications')
+  .version('0.1.0')
+
+program.addCommand(initCommand)
+program.addCommand(createCommand)
+
+program.parse(process.argv)
