+++ src/utils/inquirerUtils.ts
@@ -1,0 +1,1 @@
+export function promptUser (): void {}
